Fix bug in VaspPotcarData.from_structure() method
Initialization of the element-potential map via the from_structure()
method added VaspPotcarFile instances to the map instead of the associated
VaspPotcarData instance. This behavior is fixed in this commit and an
additional test has been implemented to assert that only objects of type
VaspPotcarData are present in the created element-potential map.