Implememt separate tests for CalculationBase class
Additional tests are implemented for the CalculationBase class
representing the parent class for all subsequent calcualtion classes.
In the added tests only the general helper methods added to the class
are explicitily tested while other methods like for instancte the
prepare_for_submission() methods should be tested in the corresponding
subclassed calculator objects.