Move submit script checks to parent CalculationBase class
General checks for the created AiiDA submit script (i.e. concerning
the use runlines and added prepend / append statements) are moved from
the subclasses to the parent CalculationBase class in this commit.
More specific checks like for the written custodian spec file which
depend on the actual calculation type are still left in the corresponding
subclasses.