Implement calculator for basic VASP calculations
The VaspBasicCalculator implemented in this commit introduced
a basic calculator object for running basic VASP calculations
(i.e. all calculations that are not of type NEB). In addition
to the regular code input the calculator plugin accepts a second
custodian code run the VASP calculation with error corrections
offered by the custodian package