Add default retrieve lists to CalculationBase
Implement default retrieve lists on the basic CalculationBase
calculation class. As the retrieval of calculation relatied
files is at the disposal of the connected parsers class **all**
generated caclulation files are retrieved to a temporary folder.
Other files, i.e. files not directly related to the VASP job,
like the submit-script, scheduler outputs and / or stdout / stderr
are retrieved and stored permanently to the retrieved folder