Implement a default parser suitable for all VASP calculations
This commit implements a basic parser class VaspFileParser that
is usuable with all VASP calculation types. The added parser class
does not attempt to parse any specific values created by the performed
calculation but simply adds the generated output files as nodes
to the calculation. In order to save on disk space the files are
stored as gzip-compressed SingleArchiveData types rather than pure
SingleFileData. Additionally, when parsing outputs generated by
NEB calculations each output file associated with a distinct image
of the NEB path is stored to an own namespace specific to that node
(i.e. a CONTCAR file belonging to the second node of the NEB path will
be stored to the output namespace 'node_02' and may be accessed via
the outputs as 'outputs.node_02.contcar')