Make custodian calculation settings storable to the database
Previous to this commit the custodian settings passed to the
Custodian code in calculation was not stored to the database
(i.e. non_db=True). However, since these settings might have
impacts on the calculated results (think of a calculation with
max_errors=1 which would fail immediately and a calculation
with max_errors=10 which would run longer, change files and
likely generates different outputs). Thus, for the sake of
data provenance this commit makes all settings passed to
the custodian code storeable