Update the serializer function for custodian.handlers input
The serializer function was updated to always return a dictionary
instead of returning either a list or dictionary. If handlers
are passed as list the corresponding dictionary with the handlers
option being left empty (indicate that default value shall be
used) is now directly generated by the serializer function instead
of leaving this task to the CustodianSettings class.