Add implemented VaspFileParser as default parser
This commit adds the VaspFileParser as default parser used
by the VaspCalculation calculation plugin.

Additionally to defining the parser itself a new dynamic output
namespace was introduced: 'parsed_results'
This namespace can be dynamically used by any parser conntected
to the calculation in order to add the parsed results to the
calculations outputs. As the concept of non-dynamical namespaces
makes sense for inputs which are **set** by the user I think it
makes no sense at all for the outputs which are not set by the user
and removes a great bit of flexibility and usability requiring
every single output that is or may ever be generated by a parser
has to be defined in the calculations spec.