Fix bug in the VaspCalculation class
Fixed bug when determining if of type NEB querying the
inputs without return a default value which would have
failed if the key was not present.

Further, queries for the submit_scrip_filename now return
_aiidasubmit.sh by default to stay compatible with AiiDA
versions below 1.2.1 where this option is not available

Tests have been updated accordingly. For one test an
additional pytest fixture has to be added to skip it
for versions below 1.2.1