Update custodian-settings and vasp-calculation sections
Needed to update the sections for the custodian settings as well
as the documentation for the VaspCalculation class due to the
changed input type for the custodian settings parameter.
The previously individually defined values have been replaced
with a dictionary version which is now also reflected in the
corresponding documentations.