Add normalization to returned Kpoint instances
When no values are defined for the kpoints-shift parameter
pymatgen defaults to a tuple of integer zeros.

This, however, does not true when values for this parameter is
present which may also be represented by float number. To unify
the values and to produced reproducible results a normalization
function was implemented transforming all kpoint-shifts to a
list of integer values