Fix SingleArchiveData compatability with file repository
Pymatgen and its VASP output classes heavily rely on the
physical existence of the corresponding files on disk.
With the new file repository introduced with aiida-core v2
the required files no longer exist physically and cannot
be used with pymatgen.

This commit introduces a fix to this issue by writing the
file contents, read from the repository, to a temporary
file on disk which is then passed on to pymatgen.