Omit mtime-bytes from comparison in unittests
During unit-testing of SingleArchiveData gzipped contents
are compared and verified against the used test data.

Currently, this comparison accounted for the complete
gzip header including the mtime stamp which lead to spurious
errors if the timestamps did not match. In this commit the
comparison is updated such that the mtime stamp is ommited
from the comparison.