Fix bug in SingleArchiveData.filepath property
Currently the _filehandle property is set during instantiation
of SingleArchiveData inside the __init__() function of the class.
However, when instantiating the class by loading a node from the
database this __init__() function is never called leading to an
AttributeError when accessing the _filehandle by calling the
filehandle property of the class.

This commit fixes this issue by removing the _filehandle property
from the __init__() function but initializing it dynamically
inside the filepath() property when it is needed