Move documentaion build to own action
Step for building the documentation was removed from the
continuous_integration action and placed into it's own
action.

In addtion the trigger for building the documentation was
adjuste slightly, such that the build is triggered for
every push but will only execute on pull requests if
file inside the docs-folder are affected.