Introduce new action for release preparation
The new action introduced in this commit is meant to
provide a basic action that is meant to help preparing
new package releases.

This action is chosen to only run on pushes to branches
names release/v** which are meant to prepare a new release
candidate. Currently, this action only verifies the versions
defined inside the package source against each other.