Implement action for automatic deployment to PyPI
In this commit a github action for automatic deployment
to PyPI is created. The workflow is triggerd if a new
release for the package is created on github

Before publishing the package, the workflow will first
check if the versions defined in the package source
match with the version tag chosen for the release. In
addition pre-commit and the unittest suite will be run.
Once all tests have passed successfully the workflow
will automatically build the package and push the
created artifacts to the PyPI.org repositories.