Implement Custodian VaspJob and ErrorHandler serializer
The herein implemented serializer functions make use of the
as_dict() function of the MSONable VaspJob and ErrorHandler
classes provided by Custodian. Instances and list of Instances
passed to the serializer functions are serialized in lists
of dictionaries with each dictionary containing the name
(i.e. the full module path) to the handler / job and the
arguments passed to the handler / job class during instantiation