Update custodian handler defaults
Previously all handler defaults have been stored in the custodian
handler defaults. This allowed to simply pass the handler name to the
calculation. However, we now require the custodian handlers to be
passed to the calculation as instantiated classes with all default
parameters already defined. This, the default have been changed to
simply update the handler parameters concerning the file names with
the defaults used by the plugin.