Introduced fixed parameter for VaspJobs and VaspNEBJobs
Since we now allow to pass VaspJobs and VaspNEBJobs directly
to the calculation, we are no longer in control of initializing
the class instanced. Although several parameters, i.e. the `vasp_cmd`
or stdout / stderr filenames, need to use the defaults used by the
plugin in order to function, users may very well set those parameters
to different parameters when initializing the VaspJob and VaspNEBJobs
class.

In order to enforce the plugin defaults this commit tags critical
arguments as fixed. Thereby, any fixed argument will be overwritten
with the defined default value, regardless of the parameter having
been set by the user previously or not.