Decouple retrieve_temporary_list from parser settings
Originally the list of temporary files, to be retrieved from
the calculation server, could be manipulated by user via the
`parse_files` setting in the calculation's parser settings.

In this commit, the ambiguity of the `parse_files` option, being
a parser option but, at the same time, affecting the retreieved
calculation outputs, is removed. This is done by introducing
a new calculation option `retrieve_files` which is purely dedicated
to affect the contents of the `retrieve_temporary_list`, defining
the set of files to be retrieved.