Introduce parser class method to signal essential files
In a previous commit the list of files that are retrieved from
a calculation was reduced from all output files to a small subset,
defined by the calculations `retrieve_files` option.

Thus, when setting up a calculation, the user would need to know
in advance which files are required by the connected parser in order
to add these files to the `retrieve_files` option. Failing to do so
would immediatle result in parsing errors due to missing files.

This commit introduces a new expected_files() method on the parser
base class. Each subclassed parser class is able to override this
method to define and return its individual list of files that are
essential to the parsing process and must be retrieved from a
calculation.