Merge essential files for connected parsers into retrieve list
This commit modifies the build process used to generate
the retrieve_temporary list. Previously, only the files
defined in the calculations `retrieve_files` option where
respected in the setup of this list.

Now, the build process respects both, the files defined
in the `retrieve_files` option as well as the files
marked as essential by the parser. This list of essential
files is returned by the parser's expected_files() method
and is merged into the `retrieve_files` list. The merged
list is then used to setup the retrieve_temporary list.