Fix bug with Custodian validators written at wrong yaml level
Before this commit Custodian validators were written on the
wrong level inside the generated yaml file. Ultimately, the
validators ended up as parameter inside the custodian settings
section

Due to this error, the calculation would fail when trying
to initialize the custodian executable from the written
custodian spec file.

In this commit, this error is corrected and the validators
sections moved to the correct level inside the yaml file