Make calculation jobs aware of custodian job suffixes
In this commit the creation of retrieval lists for both, temporary,
as well as permanently retrieved files is updated.

Up to now, only the plain files were added to the retrieve
list, leaving out any files that were extended with job suffixes
by the custodian executable. With this change, the name of each
file, marked for retrieval, is extended with the appropriate
suffix and re-added to the retrieve lists