Extend VaspFileParser with suffix handling
In this commit the VaspFileParser is fundamentally extended
and enabled to also deal with different file suffixes, that may have
been defined during the calculation setup by connected Custodian jobs.

In the original version of this parser every parsing hook, responible
for parsing a specific file, was chosen based on the filename. In order
to make this behavior work with arbitrary suffixes the hook selection
was shifted from filename based to filestem based.

To distinguish files with identical names but differing suffixes, the
linkname mechanism was updated as well. Every linkname is now prefixed
by the associated suffix (if found) in such a way, that an additional
layer of nesting is created in the outputs of the parser node. All files
with the same suffix will then be stored to the corresponding layer
defined by the suffix. Note, however, that the original behavior is
restored if no custodian jobs are defined or if no suffixes are used
for the jobs.