Fix bug when calculating the filestem
In this commit a small bug during the filestem calculation
is fixed. Previously, the filestem was calculated by simply
rstrip()ing the suffix from the filename. Since string args
passed to rstrip() are treated as character-sets rather than
actual strings this fails badly if the filestem and suffix
share the same characters.

In this commit, rstrip() is removed and replaced with an
explicit removal of the suffix based on the suffix length