Update documentation to cover new Custodian jobs options
This commit updates the aiida-cusp documentation to reflect the latest
changes in functionaly, now also allowing Custodian jobs to be defined
as input to VASP calculations.

Here, the Custodian documentation is updated as well as the
documentation of the general VaspCalculation. Moreover, the Parser
documentation is extend to cover the new parsing behavior when one
or multiple Custodian jobs are defined along with the VASP
calculation.